refactor(frontend): establish default module bundle
