+++ frontend/default/theme.js
@@ -1,0 +1,1 @@
+export const ThemeContext = React.createContext();
